Multi-Account-BTI
Every account= in .bti-config starts the definition of a new account.
Definitions before the first account= are used as default values for all
accounts. Account switch in CLI selects account. (This is indeed a
little bit ambigious and forbids two accounts with the same identity.)

This work is prerequisite to some more and serious tool for separation
of accounts. Thanks for the nice head start and relieving me on working
on the whole web interaction.

Regards,
	Lars

p.s. This is probably still a little buggy, but has been tested for
publishing.